upgrade: extract client component into /modules folder
