+++ apps/web/app/future/upgrade/page.tsx
@@ -1,14 +1,12 @@
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
 
-import { APP_NAME } from "@calcom/lib/constants";
-
-import LegacyPage from "~/maintenance/maintenance-view";
+import LegacyPage from "~/upgrade/upgrade-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
-    (t) => t("under_maintenance"),
-    (t) => t("under_maintenance_description", { appName: APP_NAME })
+    () => "Your upgrade is here",
+    () => "Improve your scheduling experience by upgrading to the new plan and enjoy the new features."
   );
 
 export default WithLayout({ getLayout: null, Page: LegacyPage })<"P">;
